feat: 어드민 사용자 행동 로그 조회 기능 추가 및 메모리 집계 리포지토리 통합
