+++ src/main/java/spring/memewikibe/infrastructure/MemeAggregationRepository.java
@@ -1,6 +1,5 @@
 package spring.memewikibe.infrastructure;
 
-import spring.memewikibe.api.controller.meme.response.MemeSimpleResponse;
 import spring.memewikibe.domain.meme.MemeAggregationResult;
 
 import java.time.Duration;
@@ -8,6 +7,4 @@ import java.util.List;
 
 public interface MemeAggregationRepository {
     List<MemeAggregationResult> findTopRatedMemesBy(Duration duration, int limit);
-
-    List<MemeSimpleResponse> findLatestMemesExcludingIds(List<Long> excludeIds, int limit);
 }
