+++ src/main/java/spring/memewikibe/infrastructure/MemeRepositoryCustom.java
@@ -1,6 +1,7 @@
 package spring.memewikibe.infrastructure;
 
 import org.springframework.data.domain.Limit;
+import spring.memewikibe.api.controller.meme.response.MemeSimpleResponse;
 import spring.memewikibe.domain.meme.Meme;
 
 import java.util.List;
@@ -9,4 +10,6 @@ public interface MemeRepositoryCustom {
     List<Meme> findByTitleOrHashtagsContainingOrderByIdDesc(String title, Limit limit);
 
     List<Meme> findByTitleOrHashtagsContainingAndIdLessThanOrderByIdDesc(String title, Long lastId, Limit limit);
+
+    List<MemeSimpleResponse> findLatestMemesExcludingIds(List<Long> excludeIds, int limit);
 }
